pyproject.toml
[project]
name = "tikz-presentations"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bibtexparser>=1.4.3",
    "dataclasses>=0.8",
    "pytikz",
    "pyyaml>=6.0.2",
]

[tool.uv.sources]
pytikz = { git = "https://github.com/allefeld/pytikz.git" }
